+++ rlapy/comps/__init__.py
@@ -1,3 +1,5 @@
 import rlapy.comps.sketchers as sketchers
 import rlapy.comps.qb as qb
 import rlapy.comps.rangefinders as rangefinders
+import rlapy.comps.itersaddle as itersaddle
+import rlapy.comps.preconditioning as preconditioning